+++ cardano-node/ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for cardano-node
 
+- Ledger peer snapshot path entry added to topology JSON parser,
+  which a new decoder function `readPeerSnapshotFile` processes
+  at startup and SIGHUP. Data is available to the diffusion layer
+  via TVar. 
+
 ## 8.2.1 -- August 2023
 
 - prevent SIGHUP from killing node during ledger replay